Use the new db-util package so we don't have to worry about specific versions
This is a significant change that will result in people getting newer version of the berk db utils package.  In order to maintain lucid support we have to check for pre-Precise and specify the old package name, but anything modern on Ubuntu / Debian can use the new package.  It just floats to the current that shipped with the release so Precise ends up with 5.1 vs. the previous 4.8.  This also allows for supporting RHEL with this package.  One step closer to RHEL support.