Updates to the slave setup with syncrepl to make it actually work
We can't assume the DN that the sync user will have.  This could be anything so make a new attribute for it
Make TLS on/off in syncrepl an attribute
Work around how chef merges attributes by not wrapping the syncrepl attributes in a slave node attribute check.  There's no need for this and it messes with the merge