Remove all the auth logic.
This auth logic actually has nothing to do with openldap and we shouldn't include it here. It's pam_ldap and should really go in some sort of pam_ldap cookbook. We will *probably* create that. We already have the newly imported to chef sssd_ldap cookbook, which is the modern way to handle ldap including sudo and password changes. Use that instead.

Signed-off-by: Tim Smith <tsmith@chef.io>