Resolve contention with apt cookbook on the mode of preseeding directory
While the mode for the preseeding directory for openldap has been `700` for many years, the apt cookbook configures it to be `755`, and when mysql configured this it used to configure it as `755`. The right thing might be to remove this particular resource, or modify it to have a specific directory per application. For now, this will solve the problem for folks using cookbooks apt and openldap, while causing a new cookbook to be downloaded with no change for other platforms.

Signed-off-by: Jennifer Davis <iennae@gmail.com>