Automatically rebuild slapd.d configuration when slapd.conf is updated
OpenLDAP 2.4 now uses a newer configuration layout where the files in
oapenldap_dir/slapd.d take precedence over slapd.conf. To workaround that, when
slapd.conf is updated, we should purge that folder and run slaptest which
regenerates the files in slapd.d as needed.

This only runs on systems that has a slapd.d directory which should only be
OpenLDAP 2.4 systems.

Signed-off-by: Lance Albertson <lance@osuosl.org>