Allow to toggle RPM repo configuration on openldap_install resource
On RHEL >= 8 the resource is configuring the OSUSOL repository to
support openldap-server install.
However in some cases you may not need this specific repository.

Now by setting `install_repository false` on the resource, the repo
won't be installed.
Also the `openldap.package_install_repository` attributes allow to
control the value set to this property in the openldap::default recipe.

Default behaviour is preserved, and RHEL >= 8 servers will continue to
install the repository.

Signed-off-by: b.courtois <b.courtois@criteo.com>